requirements.txt
aws-cdk-lib==2.149.0
constructs>=10.0.0,<11.0.0
aws-cdk.core
aws-cdk.aws_iot
aws-cdk.aws_iotsitewise